+++ src/core/settings.py
@@ -49,6 +49,7 @@ INSTALLED_APPS = (
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "modeltranslation",
+    "django_cotton",
     "home",
 )