+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## Next release
 
--
+- Implement django-cotton
 
 ## v0.2.0